backend/requirements.txt
amqp==5.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==22.2.0
Automat==22.10.0
backoff==2.2.1
beautifulsoup4==4.12.0
billiard==3.6.4.0
bs4==0.0.1
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
cssselect==1.2.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==1.2.0
Django==3.2.8
django-braces==1.15.0
django-cors-headers==3.10.0
django-csp==3.7
django-extensions==3.2.3
django-oauth-toolkit==2.2.0
django-redis==5.2.0
django-rest-framework-social-oauth2==1.1.0
django-rest-passwordreset==1.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
environ==1.0
exceptiongroup==1.1.1
filelock==3.11.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
isodate==0.6.1
itemadapter==0.8.0
itemloaders==1.0.6
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonfield==3.1.0
jwcrypto==1.4.2
kombu==5.2.4
loguru==0.6.0
lxml==4.9.2
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.0
parsel==1.7.0
platformdirs==3.1.1
pluggy==1.0.0
prompt-toolkit==3.0.38
Protego==0.2.1
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.7
PyJWT==2.2.0
pyOpenSSL==23.1.1
pyotp==2.8.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.19.1
python3-openid==3.2.0
pytz==2021.3
queuelib==1.6.2
redis==4.5.3
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
scrapfly-sdk==0.8.5
Scrapy==2.9.0
service-identity==21.1.0
simple-salesforce==1.12.3
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
soupsieve==2.4
sqlparse==0.4.2
stripe==4.2.0
tldextract==3.4.0
tomli==2.0.1
Twisted==22.10.0
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
w3lib==2.1.1
wcwidth==0.2.6
whitenoise==6.4.0
wrapt==1.15.0
zeep==4.2.1
zope.interface==6.0